+++ spec/in_measure_time_spec.rb
@@ -3,7 +3,19 @@ require_relative 'spec_helper'
 require 'fluent/plugin/in_forward'
 require 'fluent/plugin/out_stdout'
 
-describe Fluent::ForwardInput do
+describe Fluent::MeasureTimeInput do
+  before { Fluent::Test.setup }
+
+  def create_driver(conf=%[])
+    Fluent::Test::InputTestDriver.new(Fluent::MeasureTimeInput).configure(conf)
+  end
+
+  describe 'test configure' do
+    it { expect { create_driver }.not_to raise_error }
+  end
+end
+
+describe "extends Fluent::ForwardInput" do
   before { Fluent::Test.setup }
 
   def create_driver(conf=CONFIG)
@@ -40,13 +52,13 @@ describe Fluent::ForwardInput do
 
   describe 'test configure' do
     let(:config) {CONFIG + %[
-      <elapsed>
+      <measure>
         tag test
         interval 10
         hook on_message
-      </elapsed>
+      </measure>
     ]}
-    let(:subject) { driver.instance.elapsed }
+    let(:subject) { driver.instance.measure }
     its(:tag) { should == 'test' }
     its(:interval) { should == 10 }
     its(:hook) { should == 'on_message' }
@@ -54,24 +66,24 @@ describe Fluent::ForwardInput do
 
   describe 'test emit' do
     let(:config) {CONFIG + %[
-      <elapsed>
-        tag elapsed
+      <measure>
+        tag measure
         interval 1
         # hook Fluent::ForwardInput::Handler.on_read # not support inner class yet
         hook Fluent::ForwardInput.on_message
-      </elapsed>
+      </measure>
     ]}
     it 'should flush' do
       d = driver.instance
       d.__send__(:on_message, ['tag1', 0, {'a'=>1}].to_msgpack)
-      triple = d.elapsed.flush(0)
-      triple[0].should == 'elapsed'
+      triple = d.measure.flush(0)
+      triple[0].should == 'measure'
       triple[2].keys.should =~ [:num, :max, :avg]
     end
   end
 end
 
-describe Fluent::StdoutOutput do
+describe "extends Fluent::StdoutOutput" do
   before { Fluent::Test.setup }
 
   def create_driver(conf=CONFIG, tag = 'test')
@@ -85,13 +97,13 @@ describe Fluent::StdoutOutput do
 
   describe 'test configure' do
     let(:config) {CONFIG + %[
-      <elapsed>
+      <measure>
         tag test
         interval 10
         hook emit
-      </elapsed>
+      </measure>
     ]}
-    let(:subject) { driver.instance.instance_variable_get(:@elapsed) }
+    let(:subject) { driver.instance.instance_variable_get(:@measure) }
     its(:tag) { should == 'test' }
     its(:interval) { should == 10 }
     its(:hook) { should == 'emit' }
@@ -99,17 +111,17 @@ describe Fluent::StdoutOutput do
 
   describe 'test emit' do
     let(:config) {CONFIG + %[
-      <elapsed>
-        tag elapsed
+      <measure>
+        tag measure
         interval 1
         hook emit
-      </elapsed>
+      </measure>
     ]}
     it 'should flush' do
       d = driver.instance
       d.emit('tag1', Fluent::OneEventStream.new(0, {'a'=>1}), Fluent::NullOutputChain.instance)
-      triple = d.elapsed.flush(0)
-      triple[0].should == 'elapsed'
+      triple = d.measure.flush(0)
+      triple[0].should == 'measure'
       triple[2].keys.should =~ [:num, :max, :avg]
     end
   end