+++ CHANGELOG.md
@@ -1,8 +1,14 @@
+## 0.0.5  (2014/04/11)
+
+Changes:
+
+* Rename and pluginize
+
 ## 0.0.4  (2014/04/10)
 
 Fixes:
 
-* Fix to apply only plugins having <elapsed></elapsed> directive
+* Fix to apply only plugins having <measure></measure> directive
 
 ## 0.0.3  (2014/04/08)